Update student.py
STAGE 2

DECRIPTION

It's one layer of a fully connected neural network. It contains no hidden layers; all of the input-layer neurons are connected with neurons of the output layer.

To complete this stage, you will need the OneLayerNeural class that would look something like this:

class OneLayerNeural:
    def __init__(self, n_features, n_classes):
        # Initiate weights and biases using Xavier

    def forward(self, X):
        # Perform a forward step
        return ...

We suggest storing the result of the forward step in a self. variable. It will come in handy once you start working on the backpropagation.

Once you are done with the class, we need to check whether your function works. For that, create an instance of your model with the number of input neurons equal to the number of features you have, and the number of output neurons equal to the number of classes — 10. After that, apply your model to the first two items (= rows) of the training dataset and print the result.

OBJECTIVES

Objectives
Create the OneLayerNeural class and implement the __init__ and forward methods inside;

Create a class instance with the number of input neurons equal to the number of features and the number of output neurons equal to the number of classes (10);

Print the result of the model’s feedforward for the first two items of the training dataset.

EXAMPLE OUTPUT:

[0.20958963775189265, 0.8270392085632486, 0.722307754981251, 0.2541128158957566, 0.18044091327412398, 0.16085255453923492, 0.2291495292724315, 0.5164048161418893, 0.646176394989437, 0.6939396805954684, 0.311394347512693, 0.5665464164834566, 0.6121017341252446, 0.4438781312319271, 0.3350857415864084, 0.32547896896546347, 0.43611245729339715, 0.49419183444653614, 0.47815964079863255, 0.7062802881684659]